login, logout한 후 article수정/삭제, comment작성/삭제

+++ articles/views.py
@@ -1,19 +1,23 @@
 from django.shortcuts import render, redirect, get_object_or_404
     # get_object_or_404: 꺼내는데, 거기에 해당하는데이터 없으면 404status코드 제공까지 알아서해줌
 from .forms import ArticleForm, CommentForm
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse  # 응답
 from .models import Article, Comment
-# from IPython import embed  # 너무 개발용..
+# from IPython import embed  # 너무 개발용.. -> 시간을 멈춘다
 from django.views.decorators.http import require_POST, require_GET
+from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
 
 
 # render: 해당 페이지를 보여준다.
 # redirect: ex. delete하면 article 목록을 보여줄 때, index페이지(/article/)에서 보여줘야함으로 그리로 가달라고 버튼을 누르는 것!
 @require_GET
 def index(request):
+    # embed()  # request에 어떤 정보가 들어있는지 확인
     # Article 모델에 있는 모든 object를 가지고 와서 html을 통해서 페이지에 보여준다.
-   articles = Article.objects.all()  
-   context = {'articles':articles}
-   return render(request, 'articles/index.html', context)
+    articles = Article.objects.all()
+    context = {'articles':articles}
+    return render(request, 'articles/index.html', context)
 
 
 @require_GET
@@ -26,7 +30,8 @@ def detail(request, article_pk):
     # article이라는 변수에 값을 가지고오는데, Article이라는 모델에서 pk에 해당하는 정보들만 가지고와라
     article = get_object_or_404(Article, pk=article_pk)
     comments = article.comments.all()
-    context = {'article': article, 'comments': comments}
+    form = CommentForm()
+    context = {'article': article, 'comments': comments, 'form': form}
     return render(request, 'articles/detail.html', context)
         # return render(request, 'articles/detail.html', {'article': article})해도 위와 동일
         # 시험에 나올 수 있음!
@@ -37,10 +42,17 @@ def detail(request, article_pk):
 # post: 우리에게 받은 form(작성하는 페이지)에서 실제로 db에 작성(=아티클을 만들어라.))하겠다는 것
 
 
-
 # post: 내가 새로 작성을 했는데, 그걸 생성하는거 = 아티클을 만들어라.
 # get: 내가 내용을 작성할건데, 그 작성할수있는 form(페이지)를 보여줘라.
 
+@login_required  # 위치 설정
+                 # 만약 아무것도 하지않으면 기본은 '/accounts/login/'
+                 # 만약 로그인관련앱 이름이 accounts가 아니면 -> (login_urls='/users/login/')과 같이 설정해야함
+    # 바로 로그인 페이지로 보내줌
+    # url을 accounts의 로그인으로 설정 -> django는 accounts에 기본으로 setting되어있음
+    # --> /accounts/login/으로 보내줌 (이름을 설정했기때문에)
+# 로그인 끝나고나면 다시 create페이지로 보내줌
+# http://127.0.0.1:8000/accounts/login/?next=/articles/create/  => 로그인 후, 저 페이지로 보내라
 def create(request):
     if request.method == 'POST':
         # Article을 생성해달라고 하는 요청
@@ -53,8 +65,7 @@ def create(request):
             form.save()  # 데이터는 저장하는 것
             return redirect('articles:index')  # 여기서는 index에서 보여준느게 아니라 새로고침.
                             # 그러고나서, index에 들어가서 index가 보여주는 것!
-        
-    else: 
+    else:
         # GET 요청
         # Article을 생성하기 위한 페이지를 달라고 하는 요청
         # 브라우저에서 생성하고 엔터치면 나오는 페이지 보여줌
@@ -84,6 +95,9 @@ def create(request):
 #         return render(request, 'articles/create.html', context) 
 
 
+# Article 수정/삭제, Comment 생성/삭제를 사용자 로그인 상태에서만 동작 가능하도록 만든다
+# 비로그인 상태에서는 수정, 삭제할 수 없음
+@login_required  # get으로만 받는 애들을 데리고올 수 있음
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     # post의 목적: 기존의 아티클을 수정한 값을 바꾸줌 
@@ -102,45 +116,62 @@ def update(request, article_pk):
     # post: 수정하려고 데이터 들고왔구나, 원하는데로 수정해줄께. detail page로 보냄
 
 
+# @login_required  : get으로 받은 애들만 사용가능!
 @require_POST  # post요청 들어왔을 때만 가능!
                # get요청으로 들어오면 사용자가 이상한 방법으로 삭제하려고 한 것이다.
 # articles/3/delete 를 직접 입력하면 삭제됨. -> 이를 방지해야함!!!
 def delete(request, article_pk):
+    if request.user.is_authenticated:
     # article_pk에 맞는 article을 꺼낸다.
     # 삭제한다. (POST, form 사용이유: 우리가 제공하는 방법으로만 삭제하게 만들기 위해!)
-    article = get_object_or_404(Article, pk=article_pk)
-    article.delete()
-    return redirect('articles:index')
+        article = get_object_or_404(Article, pk=article_pk)
+        article.delete()
+    return redirect('articles:index')  # redirect는 get요청만 받는다.
+
 
 @require_POST
 def comment_create(request, article_pk):
-    # article 몇번인지 불러오기
-    article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        # 어떤 내용을 작성했는지, detail에서 input의 이름이 content였기때무에, request.POST.get으로 content항목을 받아온다.
-        content = request.POST.get('content')  # 사용자가 넘겨준 데이터
-        # comment라는 모델에 속하는 comment를 만들자.
-        # Comment(article=article) -> 모델에 있는 변수=지정한 aritcle이다.
-        comment = Comment(article=article)
-        # 지정한 article을 지정햇으니까 거기에 content를 넣겠다.
-        comment.content = content
-        comment.save()
-        # context = {'form': form}
+    if request.user.is_authenticated:
+        form = CommentForm(request.POST)  # 사용자가 넘겨준 데이터
+        if form.is_valid():
+            comment = form.save(commit=False)
+            comment.article_id = article_pk
+            comment.save()
+    return redirect('articles:detail', article_pk)
+
+
+# @require_POST
+# def comment_create(request, article_pk):
+#     # article 몇번인지 불러오기
+#     article = get_object_or_404(Article, pk=article_pk)
+#     if request.method == 'POST':
+#         # 어떤 내용을 작성했는지, detail에서 input의 이름이 content였기때무에, request.POST.get으로 content항목을 받아온다.
+#         content = request.POST.get('content')  # 사용자가 넘겨준 데이터
+#         # comment라는 모델에 속하는 comment를 만들자.
+#         # Comment(article=article) -> 모델에 있는 변수=지정한 aritcle이다.
+#         comment = Comment(article=article)
+#         # 지정한 article을 지정햇으니까 거기에 content를 넣겠다.
+#         comment.content = content
+#         comment.save()
+#         # context = {'form': form}
         
-        return redirect('articles:detail', article_pk)
+#         return redirect('articles:detail', article_pk)
 
 
 
 @require_POST
 def comment_delete(request, article_pk, comment_pk):
+    if request.user.is_authenticated:
     # article = get_object_or_404(Article, pk=article_pk) 쓸 필요가 없다!
     # why? urls에서 article_pk를 가지고와서 comment_delete(request, article_pk, comment_pk):에서 article_pk에 그대로 가지고옴
     #      detail페이지에 리턴할 때만 필요하므로 그 article_pk를 바로 redirect()안에 넣어줌!
     # 대신 article_pk = comment.article_id넣어줘도 괜찮. -> db보면 article_id있음
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
-    return redirect('articles:detail', article_pk)
-
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        comment.delete()
+        return redirect('articles:detail', article_pk)
+    else:  # else문 필요없음. 
+        return HttpResponse('You are Unauthorized', status=401)
+        # sessionid삭제 == 로그아웃 -> comment 삭제하기 누르면 
 
 
 # # 선생님 코드 -> form 사용
